#52 Moved interfaces into vote service
